add IndexAdapter class + move PandasIndexAdapter
